requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-simpleui==2024.1.1
Pillow==10.2.0
python-dotenv==1.0.0
PyJWT==2.8.0
mysqlclient==2.2.1
